refactor: extract the session readers from index.ts (#548 step 2b-1)
The session routes cannot move while every reader they call still lives in
index.ts — /api/sessions alone reaches for seven of them, and injecting that set
would trade one coupling for a worse one. So the readers go first: where Claude
keeps a transcript, what one says, and the sidebar rows that fall out of it.

This is the issue's step 4 item ("transcript 読み") pulled forward, because step
2's routes depend on it. The block was already contiguous in index.ts, so it
moves whole. The readers touch the registry — a live in-memory title beats the
on-disk one, and a row carries its session's activity flags — which is only
possible now that step 1 made the registry a module; the dependency runs one way.

LAST_RESPONSE_MAX moves with them and index.ts imports it rather than keeping a
second copy of the same cap.

index.ts drops 203 lines and gains nothing but imports. Verified the way step 2a
was: main and this branch booted side by side against a workspace nobody is
writing to (this session's own transcript changes under you and makes the diff
meaningless), then /api/sessions, /api/session/:id, /api/activity,
/api/transcript/timeline and /api/codex/sessions compared — identical.

Also restores a comment step 2a left behind: /api/dir-config moved to
routes/dir-routes.ts but its comment stayed in index.ts, orphaned above an
unrelated route. Observed during this step, not flagged by Codex on #553.

Refs #548
